requirements.txt
asgiref==3.3.4
decorator==4.4.2
decorators==2.0.7
Django==3.2
django-semanticui-forms==1.6.5
djangorestframework==3.12.4
hazm==0.7.0
libwapiti==0.2.1
networkx==2.5.1
nltk==3.3
numpy==1.20.2
perke==0.3.3
pytz==2021.1
scipy==1.6.2
six==1.15.0
sqlparse==0.4.1
